PlayerOne/requirements.txt
typer==0.12.3
PyYAML==6.0.1
transformers==4.41.1
torch
